requirements.txt
Django==5.1a1
django-environ==0.9.0
djangorestframework==3.14.0
django-debug-toolbar==4.4.2
drf-spectacular==0.27.2
